refactor: replace old `horizontalEvent` to new one
