+++ apps/calendar/src/components/events/horizontalEvent.tsx
@@ -1,27 +1,20 @@
-import { ComponentProps, Fragment, FunctionComponent, h } from 'preact';
-import { useRef, useState } from 'preact/hooks';
+import { FunctionComponent, h } from 'preact';
 
-import type GridWithMouse from '@src/components/dayGridCommon/gridWithMouse';
 import ResizeIcon from '@src/components/events/resizeIcon';
 import { useDrag } from '@src/components/hooks/drag';
 import Template from '@src/components/template';
-import { useDispatch } from '@src/contexts/calendarStore';
 import { cls, toPercent, toPx } from '@src/helpers/css';
-import { getGridDateIndex } from '@src/helpers/grid';
 import EventUIModel from '@src/model/eventUIModel';
-import { isNil } from '@src/utils/type';
 
 import { StyleProp } from '@t/components/common';
-import { Cells } from '@t/panel';
 
 interface Props {
   uiModel: EventUIModel;
   eventHeight: number;
   headerHeight: number;
+  isResizing?: boolean;
+  resizingWidth?: string | null;
   flat?: boolean;
-  getMousePositionData?: ComponentProps<typeof GridWithMouse>['getMousePositionData'];
-  gridColWidthMap?: string[][];
-  cells?: Cells;
 }
 
 interface StyleProps {
@@ -29,6 +22,7 @@ interface StyleProps {
   eventHeight: number;
   headerHeight: number;
   flat: boolean;
+  isResizing: boolean;
 }
 
 function getMargin(flat: boolean) {
@@ -59,6 +53,7 @@ function getEventItemStyle({
   exceedLeft,
   exceedRight,
   eventHeight,
+  isResizing,
 }: EventItemStyleParam) {
   const defaultItemStyle = {
     color: '#333',
@@ -68,6 +63,7 @@ function getEventItemStyle({
     overflow: 'hidden',
     height: eventHeight,
     lineHeight: toPx(eventHeight),
+    opacity: isResizing ? 0.5 : 1,
   };
   const margin = getMargin(flat);
 
@@ -83,7 +79,7 @@ function getEventItemStyle({
       };
 }
 
-function getStyles({ uiModel, eventHeight, headerHeight, flat }: StyleProps) {
+function getStyles({ uiModel, eventHeight, headerHeight, flat, isResizing }: StyleProps) {
   const {
     width,
     left,
@@ -111,6 +107,7 @@ function getStyles({ uiModel, eventHeight, headerHeight, flat }: StyleProps) {
     bgColor,
     borderColor,
     eventHeight,
+    isResizing,
   });
 
   const resizeIconStyle = {
@@ -125,13 +122,6 @@ function getStyles({ uiModel, eventHeight, headerHeight, flat }: StyleProps) {
   return { dayEventBlockClassName, containerStyle, eventItemStyle, resizeIconStyle };
 }
 
-function getEventColIndex(uiModel: EventUIModel, cells: Cells) {
-  const start = getGridDateIndex(uiModel.getStarts(), cells);
-  const end = getGridDateIndex(uiModel.getEnds(), cells);
-
-  return { start, end };
-}
-
 const EventItem: FunctionComponent<{
   containerStyle: StyleProp;
   eventItemStyle: StyleProp;
@@ -144,87 +134,40 @@ const EventItem: FunctionComponent<{
   </div>
 );
 
-const HorizontalEvent: FunctionComponent<Props> = ({
+export const HorizontalEvent: FunctionComponent<Props> = ({
   flat = false,
   uiModel,
   eventHeight,
   headerHeight,
-  getMousePositionData,
-  gridColWidthMap,
-  cells = [],
+  isResizing = false,
+  resizingWidth = null,
 }) => {
-  const { updateEvent } = useDispatch('calendar');
-
-  const [isResizing, setResizing] = useState(false);
-  const [resizeGuideStyle, setResizeGuideStyle] = useState<StyleProp | null>(null);
-  const lastGridColIndex = useRef<number | null>(null);
-
   const { dayEventBlockClassName, containerStyle, eventItemStyle, resizeIconStyle } = getStyles({
     uiModel,
     eventHeight,
     headerHeight,
     flat,
+    isResizing,
   });
-  const { start, end } = getEventColIndex(uiModel, cells);
-  const resetResizing = () => {
-    setResizing(false);
-    setResizeGuideStyle(null);
-    lastGridColIndex.current = null;
+
+  const { onMouseDown } = useDrag(`horizontalEvent/${uiModel.cid()}`);
+  const onResizeStart = (e: MouseEvent) => {
+    e.stopPropagation();
+    onMouseDown(e);
   };
 
-  const { onMouseDown } = useDrag('horizontal-event', {
-    onDragStart: () => {
-      lastGridColIndex.current = end;
-      setResizing(true);
-    },
-    onDrag: (e) => {
-      const mousePositionData = getMousePositionData?.(e);
-      if (!mousePositionData || !gridColWidthMap || isNil(start)) {
-        return;
-      }
-
-      const { gridX } = mousePositionData;
-      lastGridColIndex.current = gridX;
-
-      if (start <= gridX) {
-        setResizeGuideStyle({
-          width: gridColWidthMap[start][gridX],
-        });
-      }
-    },
-    onDragEnd: () => {
-      const { current: gridX } = lastGridColIndex;
-      if (!isNil(start) && !isNil(gridX) && start <= gridX && end !== gridX) {
-        const targetDate = cells[gridX];
-        updateEvent({ event: uiModel.model, eventData: { end: targetDate } });
-      }
-
-      resetResizing();
-    },
-    onPressESCKey: resetResizing,
-  });
-
   return (
-    <Fragment>
-      <EventItem
-        containerStyle={containerStyle}
-        eventItemStyle={eventItemStyle}
-        className={dayEventBlockClassName}
-      >
-        <span className={cls('weekday-event-title')}>
-          <Template template="time" model={uiModel.model} />
-        </span>
-        {flat ? null : <ResizeIcon style={resizeIconStyle} onMouseDown={onMouseDown} />}
-      </EventItem>
-      {isResizing ? (
-        <EventItem
-          containerStyle={{ ...containerStyle, ...resizeGuideStyle }}
-          eventItemStyle={eventItemStyle}
-          className={dayEventBlockClassName}
-        />
-      ) : null}
-    </Fragment>
+    <EventItem
+      containerStyle={resizingWidth ? { ...containerStyle, width: resizingWidth } : containerStyle}
+      eventItemStyle={eventItemStyle}
+      className={dayEventBlockClassName}
+    >
+      <span className={cls('weekday-event-title')}>
+        <Template template="time" model={uiModel.model} />
+      </span>
+      {flat || isResizing ? null : (
+        <ResizeIcon style={resizeIconStyle} onMouseDown={onResizeStart} />
+      )}
+    </EventItem>
   );
 };
-
-export default HorizontalEvent;
